09-04-2024

+++ nearly_lucky_numbers.cpp
@@ -3,8 +3,8 @@ using namespace std;
 int main(){
     string n;
     cin >> n;
-    int count=0;
-    for(int i=0 ; i<n.size() ; i++){
+    long long count=0;
+    for(long i=0 ; i<n.size() ; i++){
         if(n[i]=='4'||n[i]=='7'){
             count++;
         }
